fix admin can CRUD company posts and jobs conflict
