+++ app/controllers/companies_controller.rb
@@ -1,47 +1,13 @@
 class CompaniesController < ApplicationController
 	before_action :authenticate_user!, :except => [:index]
-	before_action :set_companies, :only => [:show, :edit, :update, :destroy]
+	before_action :set_companies, :only => [:show]
 
 	def index
 		@companies = Company.order('updated_at desc').page(params[:page]).per(10)
 	end
 
-	def new
-		@company = Company.new
-	end
-
-	def create
-		@company = Company.new( company_params )
-		@company.user = current_user
-		if @company.save
-			flash[:notice] = "成功建立"
-			redirect_to company_path(@company)
-		else
-			flash[:alert] = "所有欄位為必填"
-			render :new
-		end
-	end
-
 	def show
 		@jobs = Company.find(params[:id]).jobs
-	end
-
-	def edit
-	end
-
-	def update
-		if @company.update( company_params )
-			redirect_to company_path(@company)
-		else
-			render :edit
-		end
-	end
-
-	def destroy
-		@company.destroy
-
-		redirect_to companies_path
-		flash[:notice] = "成功刪除"
 	end
 
 	private
