+++ app/helpers/admin/companies_helper.rb
@@ -1,0 +1,2 @@
+module Admin::CompaniesHelper
+end
