+++ app/helpers/admin/jobs_helper.rb
@@ -1,0 +1,2 @@
+module Admin::JobsHelper
+end
