+++ app/views/admin/companies/_form.html.erb
@@ -1,5 +1,5 @@
 
-<%= form_for @company do |company| %>
+<%= form_for @company, :url => admin_companies_path do |company| %>
 
 <table class = "table table-hover">
 	<td>
