+++ app/views/admin/companies/company.html.erb
