+++ app/views/admin/companies/edit.html.erb
@@ -1,0 +1,26 @@
+
+<h1>編輯公司文章</h1>
+
+<%= form_for @company, :url => admin_company_path do |company| %>
+
+<table class = "table table-hover">
+	<td>
+		<%= company.label :title, "公司名稱"%>
+		<%= company.text_field :title %>
+	</td>		
+	<td>
+			<%= company.label :intro, "公司介紹" %>
+			<%= company.text_field :intro %>
+	</td>
+		<td>
+			<%= company.label :address, "公司地址" %>
+			<%= company.text_field :address %>
+	</td>
+	<td>
+			<%= company.submit :class => "btn btn-primary"%>
+	</td>
+<% end %>
+	
+</table>
+
+<%= link_to "回到公司文章列表", companies_path %>
