+++ app/views/admin/companies/new.html.erb
@@ -3,4 +3,4 @@
 
 <%= render :partial => "form"  %>
 
-<%= link_to "回到公司文章列表", companies_path %>
+<%= link_to "回到公司文章列表", admin_companies_path %>
