+++ app/views/admin/companies/show.html.erb
@@ -14,3 +14,4 @@
 
 
 
+<%= link_to "回到公司文章列表", admin_companies_path %>
