--- app/views/companies/edit.html.erb
@@ -1,6 +1,0 @@
-
-<h1>編輯公司文章</h1>
-
-<%= render :partial => "form"  %>
-
-<%= link_to "回到公司文章列表", companies_path %>
