--- app/views/companies/index.html.erb
@@ -1,34 +1,0 @@
-<table class = "table">
-	<h1>公司文章列表</h1>
-	<%= link_to "新增公司文章", new_company_path, :class => "btn btn-primary" %>
-</table>
-
-<div class="tab-content">
-	<div id="home" class="tab-pane fade in active">
-		<h3>公司文章</h3>
-		<table class = "table">
-			<tr>
-				<th>公司名稱</th>
-				<th>公司介紹</th>
-				<th>創建時間</th>
-				<th>最後更新時間</th>
-				<th> </th>
-				<th> </th>
-			</tr>
-			<% @companies.each do |company| %>		
-				<tr>
-					<td><%= company.title %></td>
-					<td><%= company.intro %></td>
-					<td><%= company.created_at %></td>
-					<td><%= company.updated_at %></td>
-
-					<td><%= link_to "顯示", company_path(company), :class => "btn btn-primary" %></td>
-					 <td><%= link_to "編輯", edit_company_path(company), :class => "btn btn-default" %></td>
-					<td><%= link_to "刪除", company_path(company), :method => :delete, :class => "btn btn-default" %></td>
-				</tr>	
-			<% end %>
-		</table>
-	</div>
-</div>
-
-<%= paginate @companies %>
